Move text backend from container_linux to cairo_pango
