+++ containers/linux/container_linux.h
@@ -3,8 +3,6 @@
 
 #include "../../include/litehtml.h"
 #include <cairo.h>
-#include <gtkmm.h>
-#include <pango/pangocairo.h>
 
 struct cairo_clip_box
 {
@@ -31,36 +29,14 @@ struct cairo_clip_box
 	}
 };
 
-struct cairo_font
-{
-    PangoFontDescription* font;
-	int size;
-	bool underline;
-	bool strikeout;
-    int ascent;
-    int descent;
-    int underline_thickness;
-    int underline_position;
-    int strikethrough_thickness;
-    int strikethrough_position;
-};
-
 class container_linux :	public litehtml::document_container
 {
-	typedef std::map<litehtml::string, Glib::RefPtr<Gdk::Pixbuf> >	images_map;
-
 protected:
-	cairo_surface_t*			m_temp_surface;
-	cairo_t*					m_temp_cr;
     cairo_clip_box::vector		m_clips;
 public:
-	container_linux();
-	virtual ~container_linux();
+	container_linux() = default;
+	virtual ~container_linux() = default;
 
-	litehtml::uint_ptr create_font(const char* faceName, int size, int weight, litehtml::font_style italic, unsigned int decoration, litehtml::font_metrics* fm) override;
-	void delete_font(litehtml::uint_ptr hFont) override;
-	int text_width(const char* text, litehtml::uint_ptr hFont) override;
-	void draw_text(litehtml::uint_ptr hdc, const char* text, litehtml::uint_ptr hFont, litehtml::web_color color, const litehtml::position& pos) override;
 	int pt_to_px(int pt) const override;
 	int get_default_font_size() const override;
 	const char*	get_default_font_name() const override;
@@ -82,6 +58,9 @@ public:
 
 	virtual void make_url( const char* url, const char* basepath, litehtml::string& out );
 	virtual cairo_surface_t* get_image(const std::string& url) = 0;
+	virtual double get_screen_dpi() const = 0;
+	virtual int get_screen_width() const = 0;
+	virtual int get_screen_height() const = 0;
 
 	void clear_images();
 
@@ -90,12 +69,18 @@ protected:
 	virtual void fill_ellipse(cairo_t* cr, int x, int y, int width, int height, const litehtml::web_color& color);
 	virtual void rounded_rectangle( cairo_t* cr, const litehtml::position &pos, const litehtml::border_radiuses &radius );
 
-private:
 	void apply_clip(cairo_t* cr);
+	static void set_color(cairo_t* cr, const litehtml::web_color& color)
+	{
+		cairo_set_source_rgba(cr,
+							  color.red / 255.0,
+							  color.green / 255.0,
+							  color.blue / 255.0,
+							  color.alpha / 255.0);
+	}
+private:
 
 	static void add_path_arc(cairo_t* cr, double x, double y, double rx, double ry, double a1, double a2, bool neg);
-	static void set_color(cairo_t* cr, const litehtml::web_color& color)	{ cairo_set_source_rgba(cr, color.red / 255.0, color.green / 255.0, color.blue / 255.0, color.alpha / 255.0); }
-	static cairo_surface_t* surface_from_pixbuf(const Glib::RefPtr<Gdk::Pixbuf>& bmp);
     static void draw_pixbuf(cairo_t* cr, cairo_surface_t* bmp, int x, int y, int cx, int cy);
 	static cairo_surface_t* scale_surface(cairo_surface_t* surface, int width, int height);
 };
